+++ test/BrightwayExcelWriterSpec.hs
@@ -268,6 +268,7 @@ elec =
         , activityAllocationPercent = Nothing
         , activityAllocationFormula = Nothing
         , activityNativeType = Nothing
+        , activityNativeId = Nothing
         }
 
 prodExch :: Exchange