+++ test/EcoSpold1Spec.hs
@@ -32,6 +32,7 @@ emptyActivity =
         , activityAllocationPercent = Nothing
         , activityAllocationFormula = Nothing
         , activityNativeType = Nothing
+        , activityNativeId = Nothing
         }
 
 -- | A production output exchange (isInput=False, isReference=False by default)