+++ test/ExportSpec.hs
@@ -80,3 +80,4 @@ buildFixture comp = do
             Nothing
             Nothing
             Nothing
+            Nothing